Extract `gsvg_renderer` from `gcore`, remove `gcore/vello` feature
