+++ node-graph/gcore/src/lib.rs
@@ -4,6 +4,7 @@ extern crate log;
 pub mod animation;
 pub mod blending;
 pub mod blending_nodes;
+pub mod bounds;
 pub mod color;
 pub mod consts;
 pub mod context;
