+++ node-graph/graph-craft/Cargo.toml
@@ -18,6 +18,7 @@ dyn-any = { workspace = true }
 graphene-core = { workspace = true }
 graphene-path-bool = { workspace = true }
 graphene-application-io = { workspace = true }
+graphene-svg-renderer = { workspace = true }
 
 # Workspace dependencies
 log = { workspace = true }
