+++ node-graph/gstd/src/brush.rs
@@ -2,6 +2,7 @@ use crate::raster::{empty_image, extend_image_to_bounds};
 use glam::{DAffine2, DVec2};
 use graph_craft::generic::FnNode;
 use graph_craft::proto::FutureWrapperNode;
+use graphene_core::bounds::BoundingBox;
 use graphene_core::instances::Instance;
 use graphene_core::math::bbox::{AxisAlignedBbox, Bbox};
 use graphene_core::raster::adjustments::blend_colors;
@@ -9,7 +10,6 @@ use graphene_core::raster::brush_cache::BrushCache;
 use graphene_core::raster::image::Image;
 use graphene_core::raster::{Alpha, BitmapMut, BlendMode, Color, Pixel, Sample};
 use graphene_core::raster_types::{CPU, Raster, RasterDataTable};
-use graphene_core::renderer::GraphicElementRendered;
 use graphene_core::transform::Transform;
 use graphene_core::value::ClonedNode;
 use graphene_core::vector::brush_stroke::{BrushStroke, BrushStyle};
