+++ node-graph/gstd/src/lib.rs
@@ -13,3 +13,10 @@ pub use graphene_application_io as application_io;
 pub use graphene_core::vector;
 pub use graphene_core::*;
 pub use graphene_path_bool as path_bool;
+
+/// stop gap solution until all `Quad` and `Rect` paths have been replaced with their absolute ones
+pub mod renderer {
+	pub use graphene_core::math::quad::Quad;
+	pub use graphene_core::math::rect::Rect;
+	pub use graphene_svg_renderer::*;
+}
