+++ node-graph/gsvg-renderer/src/convert_usvg_path.rs
@@ -1,6 +1,6 @@
-use crate::vector::PointId;
 use bezier_rs::{ManipulatorGroup, Subpath};
 use glam::DVec2;
+use graphene_core::vector::PointId;
 
 pub fn convert_usvg_path(path: &usvg::Path) -> Vec<Subpath<PointId>> {
 	let mut subpaths = Vec::new();
