+++ node-graph/wgpu-executor/src/lib.rs
@@ -6,6 +6,7 @@ use dyn_any::StaticType;
 use glam::UVec2;
 use graphene_application_io::{ApplicationIo, EditorApi, SurfaceHandle};
 use graphene_core::{Color, Ctx};
+pub use graphene_svg_renderer::RenderContext;
 use std::sync::Arc;
 use vello::{AaConfig, AaSupport, RenderParams, Renderer, RendererOptions, Scene};
 use wgpu::{Origin3d, SurfaceConfiguration, TextureAspect};
@@ -49,8 +50,6 @@ pub type Window = Arc<winit::window::Window>;
 unsafe impl StaticType for Surface {
 	type Static = Surface;
 }
-
-pub use graphene_core::renderer::RenderContext;
 
 impl WgpuExecutor {
 	pub async fn render_vello_scene(&self, scene: &Scene, surface: &WgpuSurface, width: u32, height: u32, context: &RenderContext, background: Color) -> Result<()> {
